Make DSA opaque
Update-Note: Accessing the DSA struct directly is no longer supported.
Use accessors instead.

Bug: 325
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/60425
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: David Benjamin <davidben@google.com>
(cherry picked from commit f4a4e2771542b90b01b15a9fe19d3fafe0fda0d0)
